Add files via upload
Not major changes in this commit.

The issue of gradient leakage into another cc when training one cc is due to Adam optimizer. Pure SGD does not result in this problem and has been tested by turning off classifier loss and checking weights of cc due to mse loss. 
When trained with SGD, training loss curves are much more jerky and it was observed once that more number of cc is converged.